Change tests workflow
